adds Jackhammer::RakeTask
I've configured a few handlers with spelling mistakes. The only way to
catch this is to wait for an exception in a live environment when the
first messages is received by Jackhammer for the handler.

Add the Rake task to your project

    # Rakefile
    require 'jackhammer/rake_task'
    Jackhammer::RakeTask.new

Run the rake task to validate your YAML will forward messages to the
appropriate handler class.

    $ rake jackhammer

It will return an exit code of 1 if any errors are found.

Jackhammer::ConfigurationValidator is an initial implementation that is
based on the YAML configuration used at RenoFi today. Main focus is to
find out if handlers are defined constants in your project. I look
forward to improvements :)