Add support for defining middleware
There is separate middleware for the client (publishing) and the server
(subscribing). The middleware supports modifying the arguments and
halting the execution.

The implementation of the `MiddlewareCollection` is fairly rudimentary,
but allows for future quality of life enhancements, such as adding code
to `prepend` middleware, `remove` selected middleware or insert
middleware directly other some other one.